rename some files and var
