--- models/user.go
@@ -1,6 +1,0 @@
-package models
-
-func CheckToken(token string) bool {
-
-	return true
-}
